Move bug10208 fix to standalone testcase
