+++ tests/PHPStan/Rules/Methods/data/bug-10208.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Bug10208;
+
+abstract class SomeParent
+{
+	/**
+	 * @return $this
+	 */
+	abstract public function clear(): self;
+}
+
+class SomeChild extends SomeParent
+{
+	public function clear(): self
+	{
+		return $this;
+	}
+}
