+++ tests/PHPStan/Rules/Methods/data/method-signature.php
@@ -481,19 +481,3 @@ class ConstantArrayClass implements ConstantArrayInterface
 		];
 	}
 }
-
-abstract class SomeParent
-{
-	/**
-	 * @return $this
-	 */
-	abstract public function clear(): self;
-}
-
-class SomeChild extends SomeParent
-{
-	public function clear(): self
-	{
-		return $this;
-	}
-}
